Updating the service providers
