+++ src/Providers/Concerns/HasConfig.php
@@ -1,0 +1,107 @@
+<?php
+
+namespace Arcanedev\Support\Providers\Concerns;
+
+use Illuminate\Support\Str;
+
+/**
+ * Trait     HasConfig
+ *
+ * @package  Arcanedev\Support\Providers\Concerns
+ * @author   ARCANEDEV <arcanedev.maroc@gmail.com>
+ */
+trait HasConfig
+{
+    /* -----------------------------------------------------------------
+     |  Properties
+     | -----------------------------------------------------------------
+     */
+
+    /**
+     * Merge multiple config files into one instance (package name as root key)
+     *
+     * @var bool
+     */
+    protected $multiConfigs = false;
+
+    /* -----------------------------------------------------------------
+     |  Main Methods
+     | -----------------------------------------------------------------
+     */
+
+    /**
+     * Get config folder.
+     *
+     * @return string
+     */
+    protected function getConfigFolder()
+    {
+        return realpath($this->getBasePath().DS.'config');
+    }
+
+    /**
+     * Get config key.
+     *
+     * @return string
+     */
+    protected function getConfigKey()
+    {
+        return Str::slug($this->package);
+    }
+
+    /**
+     * Get config file path.
+     *
+     * @return string
+     */
+    protected function getConfigFile()
+    {
+        return $this->getConfigFolder().DS."{$this->package}.php";
+    }
+
+    /**
+     * Get config file destination path.
+     *
+     * @return string
+     */
+    protected function getConfigFileDestination()
+    {
+        return config_path("{$this->package}.php");
+    }
+
+    /**
+     * Register configs.
+     *
+     * @param  string  $separator
+     */
+    protected function registerConfig($separator = '.')
+    {
+        $this->multiConfigs
+            ? $this->registerMultipleConfigs($separator)
+            : $this->mergeConfigFrom($this->getConfigFile(), $this->getConfigKey());
+    }
+
+    /**
+     * Register all package configs.
+     *
+     * @param  string  $separator
+     */
+    private function registerMultipleConfigs($separator = '.')
+    {
+        foreach (glob($this->getConfigFolder().'/*.php') as $configPath) {
+            $this->mergeConfigFrom(
+                $configPath, $this->getConfigKey().$separator.basename($configPath, '.php')
+            );
+        }
+    }
+
+    /**
+     * Publish the config file.
+     */
+    protected function publishConfig()
+    {
+        $this->publishes([
+            $this->getConfigFile() => $this->getConfigFileDestination()
+        ], 'config');
+    }
+}
